update r-build-zig to v0.6.0 and un-vendor it
Previously, we used r-build-zig-0.5.0-no-deps in order to work around
the zig build system bug that did not support http/https proxies. With
the library restructuring taking place in r-build-zig's other
dependencies, and with the upcoming 0.14.0 release which may require
further library updates, it's easier to unvendor these dependencies
for the time being.